moved utils and cleaned up logging
